+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.1.2 -
 **Changes:**
 * Fixed changelog formatting
 * Using input fields instead of placeholders in registration and login form
+* Added `validate` class to inputs
 
 v0.1.1 - 7.2.2016
 **Changes:**